Move Botkit projects inside BotBuilder libraries
