--- Botkit/Botkit.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.106
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.BotKit", "Microsoft.BotKit\Microsoft.BotKit.csproj", "{9A57B26C-9725-4A25-89FE-760B657E4F61}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.BotKit.Adapters.Slack", "Microsoft.BotKit.Adapters.Slack\Microsoft.BotKit.Adapters.Slack.csproj", "{3A464A1C-74AD-4041-A3F3-CE7C2A2FFA84}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Bot.Sample.Slack", "Microsoft.Bot.Sample.Slack\Microsoft.Bot.Sample.Slack.csproj", "{92C7525A-6B07-4246-9C71-81060EE1C021}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9A57B26C-9725-4A25-89FE-760B657E4F61}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9A57B26C-9725-4A25-89FE-760B657E4F61}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9A57B26C-9725-4A25-89FE-760B657E4F61}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9A57B26C-9725-4A25-89FE-760B657E4F61}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3A464A1C-74AD-4041-A3F3-CE7C2A2FFA84}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3A464A1C-74AD-4041-A3F3-CE7C2A2FFA84}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3A464A1C-74AD-4041-A3F3-CE7C2A2FFA84}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3A464A1C-74AD-4041-A3F3-CE7C2A2FFA84}.Release|Any CPU.Build.0 = Release|Any CPU
-		{92C7525A-6B07-4246-9C71-81060EE1C021}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{92C7525A-6B07-4246-9C71-81060EE1C021}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{92C7525A-6B07-4246-9C71-81060EE1C021}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{92C7525A-6B07-4246-9C71-81060EE1C021}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A8452F68-15BE-4C19-ABF2-1764252E8B10}
-	EndGlobalSection
-EndGlobal
