+++ Microsoft.Bot.Builder.sln
@@ -78,6 +78,12 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Bot.Builder.Funct
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "FunctionalTests", "FunctionalTests", "{8667F820-8ADA-4498-91AE-AE95DEE5227E}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.BotKit", "libraries\Microsoft.BotKit\Microsoft.BotKit.csproj", "{DB22B946-6999-4E54-8B2F-1E0C96890B6A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.BotKit.Adapters.Slack", "libraries\Microsoft.BotKit.Adapters.Slack\Microsoft.BotKit.Adapters.Slack.csproj", "{501E0B2B-F01D-4629-93E9-1AE436C66EEE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Bot.Sample.Slack", "libraries\Microsoft.Bot.Sample.Slack\Microsoft.Bot.Sample.Slack.csproj", "{6DF7693C-2A3B-4486-8257-BD068946B124}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug - NuGet Packages|Any CPU = Debug - NuGet Packages|Any CPU
@@ -341,6 +347,30 @@ Global
 		{B9DDC8CB-8EDF-4D98-913A-22F19E642223}.Documentation|Any CPU.Build.0 = Debug|Any CPU
 		{B9DDC8CB-8EDF-4D98-913A-22F19E642223}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B9DDC8CB-8EDF-4D98-913A-22F19E642223}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Debug - NuGet Packages|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Debug - NuGet Packages|Any CPU.Build.0 = Debug|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Documentation|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Documentation|Any CPU.Build.0 = Debug|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Debug - NuGet Packages|Any CPU.ActiveCfg = Debug|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Debug - NuGet Packages|Any CPU.Build.0 = Debug|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Documentation|Any CPU.ActiveCfg = Debug|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Documentation|Any CPU.Build.0 = Debug|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Debug - NuGet Packages|Any CPU.ActiveCfg = Debug|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Debug - NuGet Packages|Any CPU.Build.0 = Debug|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Documentation|Any CPU.ActiveCfg = Debug|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Documentation|Any CPU.Build.0 = Debug|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6DF7693C-2A3B-4486-8257-BD068946B124}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -381,6 +411,9 @@ Global
 		{C113E0AE-5564-4389-BA39-183A8D574210} = {AD743B78-D61F-4FBF-B620-FA83CE599A50}
 		{610963BB-2717-47AB-BFEB-C8856A4A7CA7} = {AD743B78-D61F-4FBF-B620-FA83CE599A50}
 		{B9DDC8CB-8EDF-4D98-913A-22F19E642223} = {8667F820-8ADA-4498-91AE-AE95DEE5227E}
+		{DB22B946-6999-4E54-8B2F-1E0C96890B6A} = {4269F3C3-6B42-419B-B64A-3E6DC0F1574A}
+		{501E0B2B-F01D-4629-93E9-1AE436C66EEE} = {4269F3C3-6B42-419B-B64A-3E6DC0F1574A}
+		{6DF7693C-2A3B-4486-8257-BD068946B124} = {4269F3C3-6B42-419B-B64A-3E6DC0F1574A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7173C9F3-A7F9-496E-9078-9156E35D6E16}
