+++ libraries/Microsoft.Bot.Sample.Slack/Bots/EchoBot.cs
