+++ libraries/Microsoft.Bot.Sample.Slack/ConfigurationCredentialProvider.cs
