+++ libraries/Microsoft.Bot.Sample.Slack/ConfigurationSlackAdapterOptions.cs
