+++ libraries/Microsoft.Bot.Sample.Slack/Controllers/BotController.cs
