+++ libraries/Microsoft.Bot.Sample.Slack/DeploymentTemplates/template-with-new-rg.json
