+++ libraries/Microsoft.Bot.Sample.Slack/DeploymentTemplates/template-with-preexisting-rg.json
