+++ libraries/Microsoft.Bot.Sample.Slack/Program.cs
