+++ libraries/Microsoft.Bot.Sample.Slack/README.md
