+++ libraries/Microsoft.Bot.Sample.Slack/SimpleSlackAdapterOptions.cs
