+++ libraries/Microsoft.Bot.Sample.Slack/Startup.cs
