+++ libraries/Microsoft.Bot.Sample.Slack/appsettings.Development.json
