+++ libraries/Microsoft.Bot.Sample.Slack/appsettings.json
