+++ libraries/Microsoft.BotKit.Adapters.Slack/AdapterError.cs
