+++ libraries/Microsoft.BotKit.Adapters.Slack/ISlackAdapterOptions.cs
