+++ libraries/Microsoft.BotKit.Adapters.Slack/Microsoft.BotKit.Adapters.Slack.csproj
