+++ libraries/Microsoft.BotKit.Adapters.Slack/NewSlackMessage.cs
