+++ libraries/Microsoft.BotKit.Adapters.Slack/SlackAdapter.cs
