+++ libraries/Microsoft.BotKit.Adapters.Slack/SlackBotWorker.cs
