+++ libraries/Microsoft.BotKit.Adapters.Slack/SlackEventMiddleware.cs
