+++ libraries/Microsoft.BotKit.Adapters.Slack/SlackMessage.cs
