+++ libraries/Microsoft.BotKit.Adapters.Slack/SlackMessageTypeMiddleware.cs
