+++ libraries/Microsoft.BotKit.Adapters.Slack/SlackResponse.cs
