+++ libraries/Microsoft.BotKit/BotWorker.cs
