+++ libraries/Microsoft.BotKit/BotWorkerConfiguration.cs
