+++ libraries/Microsoft.BotKit/BotkitBotFrameworkAdapter.cs
