+++ libraries/Microsoft.BotKit/BotkitConversationState.cs
