+++ libraries/Microsoft.BotKit/BotkitDialogWrapper.cs
