+++ libraries/Microsoft.BotKit/CMS/BotkitCMSHelper.cs
