+++ libraries/Microsoft.BotKit/CMS/CMSConfiguration.cs
