+++ libraries/Microsoft.BotKit/Channel.cs
