+++ libraries/Microsoft.BotKit/Conversation/BotkitConversation.cs
