+++ libraries/Microsoft.BotKit/Conversation/IBotkitConversationStep.cs
