+++ libraries/Microsoft.BotKit/Conversation/IBotkitConvoHandler.cs
