+++ libraries/Microsoft.BotKit/Conversation/IBotkitConvoTrigger.cs
