+++ libraries/Microsoft.BotKit/Conversation/IBotkitMessageTemplate.cs
