+++ libraries/Microsoft.BotKit/Core/Botkit.cs
