+++ libraries/Microsoft.BotKit/Core/BotkitConfiguration.cs
