+++ libraries/Microsoft.BotKit/Core/BotkitSlackMessage.cs
