+++ libraries/Microsoft.BotKit/Core/IBotkitHandler.cs
