+++ libraries/Microsoft.BotKit/Core/IBotkitMessage.cs
