+++ libraries/Microsoft.BotKit/Core/IBotkitPlugin.cs
