+++ libraries/Microsoft.BotKit/Core/IBotkitTrigger.cs
