+++ libraries/Microsoft.BotKit/Microsoft.BotKit.csproj
